Warn and exit when not in cache context
It is possible that the cache can be used improperly; i.e., not
in a `with cache` context. If that happened, certain methods --
namely `shrink` and `wrap_solver` -- would fail.

This commit makes it so that we don't crash in those cases.
Instead, we warn the user and do nothing.